Add logic to skip 'setup.sh' when possible
Add logic to generate a checksum file from 'package.json' and
'bower.json'. When these files don't change from build-to-build, we are
needlessly rebuilding dependencies on our Jenkins agent.

Add logic to generate a timestamp file indicating when the last
successful run of 'setup.sh' was run. We check the current timestamp
against the timestamp kept in this file, and if it hasn't exceeded a
given threshold (24 hrs), we skip doing the 'setup.sh'.

The setup logic will now only invoke when:
- the '.manifest-checksums' file does not exist
- the '.manifest-checksums' file contains mismatched checksums
- the './node_modules' dir is missing
- the './bower_components' dir is missing
- the '.last-refreshed' file does not exist
- the '.last-refreshed' file contains a timestamp older than a given
  threshold (24 hrs) when compared to the current timestamp

Tweak '.gitignore' - add '.manifest-checksums' and '.last-refreshed'