Validate the internal numbers for fallback destination
The fallback destination didn't have a validation mechanism for the
Huron numbers. It was fetching members and validating external numbers
correctly. However, when a admin wants to configure an Huron internal number,
it was so far not possible as the validation will fail. Currently, the
logic checks the backend to see if there is an absolute match available
for the huron internal numbers and enable the create button based on it.
If backend validation returns 0, then it tries for External Number validation.

Further fallback lookup data modal at the edit page was having a mismatch
and that was fixed and Unit Tests updated to reflect this change. The service
later had a minor race condition due to fetch() being called, so moved out
the fecth() call to controller. Removed all JSON formats commented code, as
we need to start refering API docs from now.