Fix the fallback userName on edit page.
The API docs says that userName field is not sent or received for
fallback destination. As a result, we need to do a API call to get
the firstname and lastname and concatinate as userName. Fixed email
field on the same card and added more unit tests for huron internal
number validation logic.