Move fallback destination validation logic to a factory
The fallback destination has a bunch of validation logic
which is needed in setup assistant controller as well as
edit controller. Moving the data logic out of controller
will help in reusing the logic. Moved out the existing
service to a services folder to keep them separate. And
updated the unit tests.