Huron US215005: Sync DIDs and DNs on changes on Number Assignment page.
Added function to automatically synchronize the DIDs and DNs when
selections change on the Number Assignment page.

Also updated the grid for the Number Assignment so that calls to
DialPlanService are not triggered unless the users have UC entitlments.

Also fixed an issue where a .catch wasn't being triggered correctly
and at the right time when errors were encountered fetching customer
data in serviceSetup.controller.js.