Fix the edit page load async behavior
The members lookup and the fallback data are fetched asynchronously
right after the hunt group get is completed. This causes the edit
page to load first before the hunt members are loaded. And as a
result there is a short duration in which a warning is shown about
members unavailable. Added promise synchronization for all async
tasks during edit page loading.