Add invalid input indicator for HG typeahead control
The invalid input indicator underlines the input text with a
red line if the data is found invalid. This is implemented in
toolkit and reused here. The hunt pilot number, hunt members
and fallback destination input will have this error indicator.
The logic kicks in after 3 key strokes and underlines in red
if the input entered is invalid.