fix(tool): point default 'atlas_api_url' at integration
As 'curl-atlas' is a tool for helping developers, it is more likely
that it should be pointing at Atlas in integration by default. Set the
environment variable to point to integration by default, but allow
overriding it by taking any predefined value for 'atlas_api_url' if
one is already defined.

doc(tool): fix typo on usage doc