Hardcoding Huron models
This means adding a switch for getting a human readable model name, and a switch for getting the image for the supported types. This eliminates the need for the detailed Huron API calls, making the right hand pane faster to load!

Keeping some code for periodically loading additional details on a device commented out, we will need it for fetching lines for Huron devices soon.