Replacing flash based copy to clipboard with a javascript based one
This gets us around problems with Flash support, configuration and browser permissions. The plugin we use now (https://clipboardjs.com/) is not supported in Safari but offers the possibility to display a message when copying did not work. In a later commit we will use that to select the text that was being copied and tell the user to "Press Ctrl+C".