fix(Core): hide first subscription header only when it has trials
Previous logic assumed that there could ever only be:
- 1 subscription (of active trials)
- 1 subscription (converted)

However, this turns out not to be the case, as a customer org can
make multiple conversions, which results in multiple subscriptions
(ie. multiple rows appearing in the 'License Usage' tab).

This change makes the first subscription's header hidden only when
it happens to contain an active trial. In all other cases it is
fine to show the header.

refactor(Core): cleanup double-quotes