US213389: UI Template for open/Closed Schedule with Message, Phone, Menu.
Added error handling for deleting a predefined schedule when there is
a CE save failure.

Added error handling for staying on the AA Name page when there is a
CE save failure.

Define a new flag isAANameDefined (replacing ui.ceInfo.name) to be used
in html to determine if an AA Name is defined because saveAARecord
always initializes ui.ceInfo.name, which unhides the overlay panel upon save.