GHI 318 - EFT - (CDW) Todo notice appear after customer has completed setup Call Service Aware&Connect
The state checker service defaultet to showing a notification if the reply from USS was undefined. This yielded an errounous notification when the admin first loads the page, and the reply (from the cached value) has not yet returned -> we would show a notification that some users are not enabled for Aware/Connect erronously.

Fixed by inverting condition -> only showing notification if reply from USS is not undefined.