Fixing SIP URI notification never removed in Hybrid Services
The wrong service (USSService2) was used to get the orgSettings that we check whether the SIP Cloud Domain is set or not.
Now changed to using the OrgService to check if value is set.
Also added an in memory ACK to the nofification (will be loaded again if the user logs out and in again). This ack has the same look that is used for notifications in the Overview page.