atlas-466 - Call Service status reflects only Aware
Fixes https://cibu-jira.cisco.com:8443/browse/ATLAS-466
When doing a review, see test text: Now the following is shown as an aggregation when admin clicks a user:
- Aware: entitled, Connect: not entitlet => follow Aware status
- Aware: entitled, Connect: entitled, error => error
- Aware: error: Connect, entitled => error

etc.

The only thing I am slightly not sure about, is the order of significance, which is now:
- not entitled
- activated
- pending activation
- error

That is, if one entitlement is activated ok, but say Connect is pending, we will now show 'pending actication" instead of "Activated".  This can be discussed, I guess.