Fixing ATLAS-396.
We now parse the data we get from certs-integration.wbx2.com, and display it in a table instead of as a verbatim string.

Created a new service to do the formatting, so that we do not pollute the existing certificate service.

Creating unit tests for the new service, and a unit test in the controller that verifies that we use the new service correctly.