add working links to Reports icon and Site Config icon in Actions Column
remove references to custom style (color of icons) in .tpl.html, but
left customer style definition in .scss for the time-being.   Get
comments and approval from UX team on default behavior.

changed e2e files for sitelist, reports, sitesettings to go through new
codes in the Actions Column, but leaving the old codes intact that test
by going through the individual 'Reports' column and the 'Configure'
column.

When we are ready to remove the code for the individual 'Reports' col
and 'Configure' col, we will then remove rest of codes associated with
the individual-col implementation.