DE2979: Ensure that test cases remove Atlas_Test_UI*
When the integration-partner_spec end-to-end test is run, it creates an
organization for testing purposes, but does not delete the test
organization.  Previously, in order to get rid of the test organization,
there was a manual process that had to be done using a shell script.
This commit will address that concern by automatically removing the test
organization upon successfully running through previous steps in the
end to end test using the Delete button that was recently introduced
into the frontend.  There is, however, still the chance that due to
network instability, this test case will fail and will require manually
removing the test organizations, though the likelyhood of this is rare.