Global Settings: Including branding section, with modification of branding
The new and old branding is added as components, which included in the branding component.
There is a new BrandingFeatureController with a single task of quering the feature toggle service which is used by the branding component.

Definition of which controller to be used is moved from the templates and to the components and usage of the templates ( example in partnerProfile.tpl.html)