Add total user validation and e2e test
Total user licenses should not exceed care license count. Add a
basic happy path flow that is able to create a trial with care
offering and find's the trial on the customers page. After
finishing the suite delete the org. The test reuses the
partner-admin org for creating trial.

Add total user validation with care