fixing gulp watch to not copy javascript spec files on spec file change
The problem was that when you changed a javascript spec file, the gulp watch task would copy those files to the build directory.
Test files are run from the original location thus it ended up duplicating the test files.

Fixed by filtering the spec files in the sub copy task in the gulp watch task