US12857: Front End - In-service Retention Policy
This user story involves implementing the new data retention endpoint in
the global settings page. This endpoint differs from the usual endpoint
that is used by the first-time wizard since the first-time wizard should
not be revisited more than once (hence why it's a first-time wizard).

When a user selects a retention policy that is lower than the selected
one, a modal window will pop-up and alert the user that the desired
changes will remove messages older than the retention policy selected.
If the user cancels the changes, the dropdown will revert the changes.

The retention changes will go in by AJAX - there is no requirement to
click on a Save button.

I have also taken the time to rename "dataPolicy" to "retention"
throughout the global settings page, for better readability.