Patch first time admin login with Care
The customer org admin needs SyncKms role and Care
related entitlements to do Care agent on-boarding and
any other Care related admin operations. During trial
creation, the Org will be given the entitlements and
required scopes. However, the Org Admin will be given
only the free entitlements and default hybrid service
admin entitlements, since these operations are done as
Partner admin.

Until we figure out a better way of patching the new org
admin's scope and entitlements, this is a work around to
patch the admin when he logs in for the first time into
Atlas UI. Further, he will be forcefully logged out during
the first time login, to enable him to re-login with the
newly patched entitlements. Ideal way is to revoke and get
a new access toke on the fly which will be pursued shortly.