US12944 Added expired badge to expired trials in customer overview
In the customer overview page expired licenses now have an 'expired' badge instead of a trial badge.
This feature is also in the License Usage tab as well.

Note that some functionality (getting the number of trial days left) was moved from
the trial notice banner controller to the TrialService.