Adding new Device Activation Wizard
The new wizard will be used to create activation codes for huron and
cloudberry devices from both the users page, places page and device
page in atlas.

The wizard can be configured to use different flows depending on the
context it is used. But each step will have the same behaviour regardles.

So for example when you add a device from the places page you will not
be asked if it is for a user or a place, as this is implicitly given.

Most of the wizard steps is hidden behind a feature toggle, as we do not
yet support adding huron devices to shared spaces.

The change that will be noticable to a customer will be that it is now
possible to add a huron device for an existing user from the device page.