Hide global settings behind atlas-settings-page feature toggle
The global settings page is accessible by all organizations.  This
commit will address that concern by putting the page behind the
atlas-settings-page feature toggle.  This means that in order for an
organization to be able to access the global settings, they will have to
have the atlas-settings-page feature toggle enabled.