US12874: Move the branding logic in global settings page
If the user is not a partner admin and they are not a partner user, then
not _every_ user is a direct customer.  The correct way of checking if a
user is a direct customer is to check the commerce relationship for
whether or not it contains the appropriate value.