Fixing search to allow specifying entitlements
This was in place at some point in the past but stopped working due to refactoring it looks like. We added the entitlements parameter back in and fixed the filter string to work with this.

The filter string had an issue we had to address, it needed parentheses to make the 'or' statements be optional. Before our fix, the 'active eq true' was never actually enforced because of the 'or' between the name requirements.