CSDM: Added initial draft of Places Page
This will be the page to view and configure places, to which devices can be assigned. Both huron (desk phone) and cloudberry (room system) devices will be part of places. At the moment all of this is behind feature toggle 'enablePlaces'. This feature is still under development but we need it to run in master to see how it behaves.