US12878: Surface Status From Mailgun for Emails Sent From Atlas - Helpdesk UI
When emails are not received by end users there is no way for the tenant
admin to figure out why this is happening.  Thus in every case the
tenant admin has to contact the TAC.  TAC also has not way to see
information about the email status so they send the request straight to
engineering.  As a result a high number of escalations are being sent to
engineering.
There should be away a way surface email status in  Atlas and Help Desk
tool so that we can at least reduce the number of escalations to our
developers.