docs(github): adds issue and pull request templates
This pull request introduces GitHub templates to the Atlas UI.  When
people try and create a pull request or open a new issue, they will be
shown the text contained in these templates.  By having people manually
check off a box that confirms they wrote a unit test and ran the
testing, there is a good chance that half the build errors we get due to
static analysis or unit testing will be resolved.