fix: Correct the config to use default modal rather than small
A change in Atlas exposed a bug in where we used the wrong modal for our add device flow. At the time we pushed this flow it used the default size modals, but recently it has switched to the small modals. Our design has the default size modals so that's what should be in Atlas.