fix(helpdesk): Added new url to the readonly interceptor to avoid toaster when in readonly launch mo
Some code have been added to run a DELETE request on user tokens when signing out from Atlas. This

caused the readonly interceptor to stop the request when the user is logged in as a readonly admin

user. We want there request to pass even in readonly mode. This will also avoid getting the

interceptor toaster on signout.