fix(hercules): Fixing bug where Hybrid Calendar setting was displayed also for other Hybrid Services
A new section for setting whether or not to send emails to end users was introduced by the Calendar
team some days ago. The setting was placed in a template that is also used for other Hybrid
Services, so that the CAlendar specific setting also was there in the Hybrid Call and Management
Service pages.

https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-1009
https://sqbu-github.cisco.com/WebExSquared/wx2-admin-web-client/pull/2330