fix(hercules): Changing the Hybrid Calendar user email setting.
Updating typography and strings after discussing the new feature with our UX designer Jacopo, and adding localization support.
Also moving error handling code from service to controller, as it is the usual pattern in the Hercules module.
On save errors, do not flip the switch, but put it back to its original position.

https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-1009