feat(core): Adding BMMP to npm, including it in the my subscriptions page, and adding a typings file
Adding BMMP to atlas via npm and creating a typings .d.ts file that is referenced in the

typings.json file to allow BMMP to be easily used anywhere in Atlas.

Adding the BMMP upgrade/change

plan widgets to the My Subscriptions page, but keeping it hidden as the current version redirects

all subscriptions to '#'; the current BMMP workaround remains in place.

Refactoring some of the

mySubscriptions code for readability.

US13033