feat(hercules): Adding upgrade schedule to cluster sidepanel
As part of F410, clusters get an upgrade schedule. Our UX wireframes show the schedule in the
cluster sidepanel, but this was for some reason never implemented. As part of the fix, I'm also
moving the functions for localizing the date / time from the cluster list controller to
FusionClusterService, so that I do not have to reinvent the wheel.