fix(Hercules): Cleaning up Hybrid Service deactivation
Changes: Aligning the text and visual design of the deactivate service dialog with the deprovision service dialog. Changing the color of the "Deactivate" button in the Hybrid Call & Calendar pages. Updating the help text in the Hybrid Call & Calendar pages. Redirecting the admin to the services overview page (not the overview page) after deactivation.

https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-1112