feat(core): Display info when organization customer does not exist
This should only ever happen in testing. If a user logs in to an organization
defined on the Integration server from a Production server backed UI, then
they will not have access to editing Services for the users.  This manifests
by displaying a blank screen in the User->Services->Edit Services screen.

This change adds some feedback to let the user know this is not a bug and that
they need to verify they are connecting to the correct backend.

The problem stems from the call to /customers?orgId= not finding the orgnization
and returning an empty customers list.

ATLAS-1116