fix(core): Skip Assign Lines page when pressing Back in Add Users
In the first-time setup wizard, if no Call service is selected, flow correctly skips the Assign

Lines page.  However, after adding users, if the customer presses the Back button, they are taken to

the Assign Lines page instead of the Assign Services page.  This fix adds a check for this and now

correctly skips the Assign Lines page when no Call service is selected.

ATLAS-1200