fix(core): Remove invalid text on FTW add-user summary and extra Skip buttons
When Adding Users in the FTW, there was incorrect text on the summary screen "Select a method to
upload your users".  This fix removes that text.  Also, the Skip button will now only appear on the
first page of Add Users since it is confusing on the following pages.

ATLAS-1199