fix(devices): Make sure all devices and places are loaded before presenting the places data model.
There was a bug where the order of api calls determined how many huron places with devices we got.

This was an optimization to be able to display the list faster, but too buggy. The solution was to

make sure we always wait until everything is loaded before displaying the places list. We can

revisit to optimize later, but first make it right!