fix(hercules): no duplicate alarms on cluster overview panel
Given that all the nodes of a cluster share the same configuration,
when one node has an issue, it is very likely that all nodes will also
have that same issue.  This leads to alarms being raised by each and
every node, and, more often then not, the alarm has the exact same
content on each node.  Currently, Atlas shows the concatenation of all
alarms on all nodes, so ends up showing the same alarm contents
multiple times.  The changes in this commit make the aggregation code
eliminate duplicates.  When several nodes of a cluster emit alarms
that have the same ID and contents, we collapse them into one single
alarm, thus reducing the cognitive burden for the administrator.  It
will still be possible to see per-node alarms (and so see that an
alarm is present on multiple nodes), but the cluster overview panel
will not have duplicates anymore.

For example, given a cluster of 4 nodes running the Call connector:
since all nodes share the same configuration, and thus the same
on-premise UCM server address, if one node can't talk to the UCM
server, all nodes can't talk to the UCM cluster.  This leads to 4
alarms reported up by FMS - one for each of the node - about not being
able to communicate with UCM.  Before this commit, Atlas would show
this alarm 4 times in the overview of the cluster; after this commit,
it will only show it once.