feat(core): Changing 'Three Months' to 'Twelve Weeks' for Most Active Users table.
Under the feature toggle for the 3+ time filter, changing Most Active Users table title and

descriptions to use 'twelve weeks' instead of 'three months' for increased accuracy.

Switching the

new API for Most Active Users out from under the feature toggle since there is no design impact in

doing so and was only kept in the feature toggle due to an initial lack of data that has since been

remedied.

Updating the customer reports page to use only one type of masonry resize and updating how

the resize is called to minimize overlap between the Active User card and the card beneath it.

US14797