feat(devices): trial support for darling
Trial support for darling.
UX where you specify number of spark boards in the trial add and edit dialog
this is hidden behind the atlasDarling feature toggle
and adds the SB offer to license count for devices in the customer overview for partners.
with the option to drill down to the different licenses.

US14448