fix(core): Fixes (...) menu for 2nd to last user in Users List so that it doesn't vanish.
The (...) menu for the 2nd to last user in the Users List would appear below the (...) and then

disappear from view entirely.  Even when it doesn't disappear, if there were multiple entries in the

menu, then some would be cut off at the bottom.  Used :nth-last-of-type to force the menu to appear

above the (...) instea of below, which fixes both the disappearing act and the cut off options.

ATLAS-1351