fix(hercules): Incorrect warning in the cluster “Upgrade Now” section
The logic for calculating whether or not to display the “This cluster has one or more nodes where Management Connector is not sending updates to Cisco Collaboration Cloud. It is recommended to bring Management Connector online before initiating an upgrade.” warning has been incorrect since July(!). Fixing it, and adding another test which asserts that a healthy Cisco alpha cluster does not trigger the warning.

https://sqbu-github.cisco.com/WebExFusion/fusion-alpha-issues/issues/782