fix(squared): Handle error result in healthMonitor call
ReportsService.healthMonitor uses a callback to return results of  request. However, if there
is an error then no data is returned.  In this case, the callback functions would check the
'data.success' flag, but throw a runtime error because 'data' is undefined.  This fix now ensures
data.success is defined on .errors.

ATLAS-1446