fix(devices): Update currentDevice with new upgrade channel
The device was reloaded and the cache updated but the currentDevice on the overview page is not linked to the cache. Setting it now when the device is reloaded for checking the upgrade channel. Also fixed the error message when the device doesn't update in time, that did not show the error text.
See https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-1555