fix(hercules): Use serial instead of hostname to identify node in node details sidepanel
Our code for expanding a node in a cluster had the expectation that hostnames always are unique within a cluster. They aren't necessarily so, if a node dies and you recreate it using the same hostname, you can have two nodes (one alive and one offline) with the same hostname in the same cluster.
Now changing the sidepanel code to rely upon serials instead, as they should be unique even in such cases as described above. Also fixing the sort order of connectors so that you get a determenistic order if you have multiple nodes with the same hostname.

https://sqbu-github.cisco.com/WebExFusion/fusion-alpha-issues/issues/866