feat(core): userOverview no longer modifies the  userData
userOverview was modifying the .currentUser.  This was causing ui-router to reload the
entire UI state whenever this change happened.  Also, it is bad practice to modify the stateParam
since they should be considered read-only.

This change introduces a userOverview.service that makes the neccessary modifications to currentUser
in the resolve for the user-overview state, before switching to the new state.

 - changes from code review
 - work towards untangling from global module definitions

ATLAS-1559