fix(Hercules): Resource Group “Remove” button not refreshing correctly.
When I unassigned the last cluster from a Resource Group, we didn't enable the “Remove” button. Likewise, when I added the first cluster to the Resource Group, we didn't *dis*able the button. Fixed it by adding an extra check for whether or not the button should be enabled whenever the “Assign Clusters” modal is saved.

https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-1674