refactor(trials): US6033 - move trial email sending from client to backend
This is the last change in a series that removes email sending from the client.  Instead, email will
be sent via a new, dedicated backend endpoint, or as part of an existing process (e.g., create
trial).
This was being done behind a feature toggle, but is now enabled by default so all related
code can be removed from the client (which actually means also removing the EmailService)

[US6022](https://rally1.rallydev.com/#/76458540020d/detail/userstory/76464760624?fdp=true)