fix(trials): search filter not always filtering results
In the automated tests, the customer search filter is being populated with text, but the results are
not filtered with it.  This may be a race condition between the loading of data in the table, and
the entering of the text.  I've tried to fix this with both a change to the customerList and also a
change to the test utils.
Also made a code change to suppress the 'autotrack' traces in the karam
tests (caused by mixpanel).