fix(hercules): Notifications were not calculated in the Hybrid Call / Calendar settings pages.
The component that creates the service specific cluster list was reposible for subscribing to and updating notifications. They were thus never loaded if the user only visited the settings page. Now moving the logic for subscribing to and updating notifications to the expressway common container, which also is loaded by the settings page.