fix(core): Fixing left-nav so that child pages still count as selected
When a child page is open - such as care reports - the associated tab does not appear as selected;

only the main page causes the tab to display as selected.  Adding a check to verify where the the

current location is a child page based on the path such that '/reports/care' will still have

'Reports' show up as selected in the left nav, just like the main page - '/reports' - will.

ATLAS-1332 - https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-1332