feat(devices): Rework activation code wizards for personal Cloudberry
Changing the wizards to work with the idea of one activation code. Right now CMI is still used to create Huron personal activation code when the 'ce-enable-personal-mode' feature toggle is not set, otherwise CSDM is used for all personal codes. That will not work on Huron phones until we start using the new combined CI config id so we won't merge this PR until that's ready.