fix(core): Manage Users from Users Card on Overiew now works
CLicking the Manage Users button in the Users Card on the Overview tab would bring up the Manage
Users dialog, but it would not work correctly.  This was because Manage Users requires some setup
from the Users List page. Now forces transition to the Users List before opening the Manage Users
dialog.

ATLAS-1659