feat(Pickup):  Gray out and disable members already in a pickup when being selected
Requires:  collab-ui-ng 3.4.0
This feature will gray out and disable a member from the select screen if all lines are in a pickup group.
If there is an availabile line, it will grey out and disable lines from selected after selecting the member
and display a warning saying the user is already in another pickup group.

US261787 - UI: Call Pickup Wizard - Members Page -- 2 of 2
1) Admin selects the Number they want to add to the Pickup Group.
2) If the number is assigned to a Pickup Group, display warning dialog box

https://rally1.rallydev.com/#/14990096206ud/dashboard?qdp=%2Fdetail%2Fuserstory%2F75674747476