fix(core): Fix for failing manageusers e2e tests
Removed activating users from manageusers-*spec.js since we don't need to activate the users in
order to determine if they were onboarded.  There is an intermittant error that is occuring when
activating a users and setting a password in the e2e tests here.  This change should bypass that
issue by no longer trying to activate a user.

 - removed id for 'Finish" button in addUsersResults modal since it isn't needed.
 - test to ensure Free/Paid services are correctly enabled

ATLAS-1776