fix(Pickup):  Disabled enter key from selecting a disabled user and added disabled member CSS to edit page
Bug:  When a user was disabled you could select by using the enter or tab key when you shouldn't be able to.

Fix:  This fix was tricky because ng-collab-ui would set global hotkey listeners.
I turn off the hotkey keydown event listener and create my own as a directive and added back the needed hotkeys.
When the keyCode detected is tab or enter, we check if the member is disabled.
If the member is disabled, we block the key event.
The rest of the code is to act the same as hotkey listener like in ng-collab-ui.

Bug:  When a user has all lines taken, it should grey out the user on the edit page.
Fix:  I added CSS to grey out the disabled user.

Original US: US261787

JIRA: https://jira-eng-chn-sjc1.cisco.com/jira/browse/PROVISION-3273