fix(login): allow login transition for any successful health status response
Currently if `HealthService.getHealthStatus()` resolves any value other
than `online`, then there is no state redirection and the admin is left
on a blank page.  The admin should not be left unknowingly in a bad
state, nor should they be redirected to a "Server Maintenance" page
when the status is in fault.