fix(helpdesk): DE857 - user language change
If you open the Help Desk tool and try to change the user language, the page reloads and sends you
to the customer portal (at least for a user that has roles where that would be allowed). If you then
try to open the Help Desk tool again (clicking the button in the Support page), the language is back
to English. Expected behavior would be:
After the language change, you stay put in Help Desk
If you launch Help Desk from the Support page, the current language selection should be sticky.

https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-1597