fix(devices): Warning on Launch Advanced Settings for read only admins
The button currently opens a new tab that tries to launch to the endpoint but fails because the read only admin fails to send a mercury message to the device. Going back to the Atlas tab shows the user that this is because he is in read only mode. The button should not launch the new tab and just show that warning. See https://jira-eng-gpk2.cisco.com/jira/browse/SQRD-1109