fix(core): Ignore IsOnline org for Pending User status
Change the criteria for determining if a user is Pending or Active.  Previously, it would show the
user as Active if they had an entitlement and were in an Online org.  Now, it will only show Active
if the user has an entitlement AND either they have signed in OR have the 'ciscouc' entitlement.

ATLAS-1868