fix(core): ATLAS-1320 - My Subscription displays incorrectly for non-CI site
The License Summary on the Subscriptions page should not display the usage for a non-CI site and
provide a link to where the non-CI site's licenses can be managed.
Refactoring mySubscriptions to be
more readable and follow typscript conventions better.
Fixing css/class behavior so that the new
appearance behind the feature toggle doesn't alter the current appearance when the feature toggle is
false.

https://jira-eng-chn-sjc1.cisco.com/jira/login.jsp?os_destination=%2Fbrowse%2FATLAS-1320