feat(places): Make it possible to set Software Channel on a place
First cut for this feature, for now behind toggle 'ce-enable-lyra-config'.
A place will show a new section "Device Settings" with currently only "Software Upgrade Channel". Initially this will be set to "Nothing Selected", which means the device's local setting for upgrade channel is used. If a device in the place does not yet support reading the setting from the cloud, a warning will appear below the dropdown (currently no device supports this yet!). The setting is managed in Lyra, Atlas reads it from and writes it to Lyra.
Removed CsdmConfigService to make this change easier.
Next steps:
- Unit tests
- Use the same component on the Users page (might have to make the Users page first also use the ucDeviceList component like the Places page does)
- A lot of UX attention :)
- Changes on the device fly-in based on this setting