feat(core): Add arrow indicating more content to scroll to
In the FTSW, some panels have more content then can be displayed vertically.  This requires the user
to scroll the panel to see more content. However, there is no indication that there is more content
to scroll to since the scroll-bars on most browsers are now hidden by default.  This feature adds a
"down arrow" indicator at the bottom of the content panel to indicate there is more content below.
Also, if the user clicks the down-arrow they are taken to the bottom of the content.

 - added attribute directive 'scrollIndicator' that will add a down-arrow button to a container if there is more content to be scrolled in to view.
 - fixes to scss to better support flex
 - change to _common.scss so that all [ng-click] use cursor:pointer

ATLAS-1733