fix(core): US15271 Do no cache My Subscription DB query
The DB query that is used to populate the My Subscription page is cached, which is causing a problem
for Online orders. After doing a Change Plan to purchase new orders the user has to log out of Atlas
and back in to see the updated order info on My Subscription.
This PR provides a new parameter for
the query to specify whether or not to use the cache (the method is used by other pages) and also
refactors the existing code so that the query is only called one.

US15271 Refactor My Subscriptions page so it updates without having to log out