fix(core): Remove vestigial LeaderBoard from main page
LeaderBoardCtrl and associated directive are no longer used since subscription/license data is
displayed in the My Company page.  The directive was still being injected in main.tpl.html, taking
up a space in the Tab order.  This fix removes the entire leaderBoard directive and controller and
excises it from the main html.

[ATLAS-1958](https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-1958)